First Commit
This is first prototype of the game. Very early stages.
The overall design is a little akward and my custom controls could use
some serious updates.
The game is not in a playable mode but it should run and allow you to
choose a faction and let you build ships, facilities, and troops. You
can also move ships.
The data loads from XML files and the images loaded are not mine and are
intended only for use in prototyping this project.
The intent of the game is to allow players to add their own content and
manipulate the game themselves.